Add support for passing audience token through cmdline
This commit adds support for passing audience token
through cmdline arguements of the sidecar.
This avoids additional API call required to fetch
the SnapshotMetadataService CR to obtain audience
token.
If audience token is not set in cmdline args, the
previous logic of fetching audience token from
SnapshotMetadataService will be used.

Signed-off-by: Rakshith R <rar@redhat.com>