Parse go version from release-tools
trivy.yaml should use the same go version as release tools, so it checks
for CVEs with the same go as we use to build images.

Add a step that parses the go version from prow.sh and stores it in
goversion file. Use the file in subsequent setup-go step.